requirements.txt
# -*- coding: utf-8 -*-

Django==2.2.6
requests==2.22.0
celery==4.4.7
redis==3.2.0
redis-py-cluster==2.1.0
django-redis==4.8.0
django-celery-beat==2.2.0
django-celery-results==2.0.1

pytz==2019.3
gunicorn==19.6.0
six==1.11.0
urllib3==1.24.3
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-jsonfield-backport==1.0.3
uWSGI==2.0.17.1
python-dateutil==2.5.3
djangorestframework==3.12.2
django-filter==2.1.0
arrow==0.14.2
whitenoise==3.3.0
arrow==0.14.2
mistune==0.8.4
gevent==1.2.2
python-consul==1.1.0
PyYAML==5.3
dataclasses==0.7

# for原生es
elasticsearch==7.0.0
elasticsearch5==5.5.6
elasticsearch6==6.4.2
elasticsearch_dsl==7.0.0

# kafka
kafka-python==1.4.6

# iam
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
curlify==2.2.1
idna==2.8

# qcloud cos
cos-python-sdk-v5==1.9.0

# pipeline
bamboo-pipeline==3.4.0
mock==2.0.0

# jinjia
jinja2==2.10.1

# 各自环境所需的不同依赖, 通过 sites/${env}/deploy/requirements_env.txt得到
-r requirements_env.txt
